update duration in material

+++ EduQuest_Application/DTO/Request/Materials/AssignmentRequest.cs
@@ -5,7 +5,7 @@ namespace EduQuest_Application.DTO.Request.Materials
 {
 	public class AssignmentRequest : IMapFrom<Assignment>, IMapTo<Assignment>
 	{
-		public int? TimeLimit { get; set; }
+		public double? TimeLimit { get; set; }
 		public string? Question { get; set; }
 		public string? AnswerLanguage { get; set; }
 		public string? ExpectedAnswer { get; set; }
