+++ EduQuest_Domain/Entities/Material.cs
@@ -12,7 +12,7 @@ namespace EduQuest_Domain.Entities
 		public string Description { get; set; }
         public int? Version { get; set; }
         public string? OriginalMaterialId { get; set; }
-        public int? Duration { get; set; }
+        public double? Duration { get; set; }
         public string UserId { get; set; }
         
         //Video
